LayoutTests/inspector/syntax-highlight-html.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="syntax-highlight.js"></script>
<script>

// Frontend functions.

function frontend_dumpSyntaxHighlightHTML(str)
{
    return frontend_dumpSyntaxHighlight(str, "text/html");
}

function frontend_doitAndDump() {
    var result = [];
    result.push(frontend_dumpSyntaxHighlightHTML("<html>"));
    result.push(frontend_dumpSyntaxHighlightHTML("<table cellspacing=0>"));
    result.push(frontend_dumpSyntaxHighlightHTML("<input checked value=\"foo\">"));
    result.push(frontend_dumpSyntaxHighlightHTML("<table cellspacing=\"0\" cellpadding='0'>"));
    result.push(frontend_dumpSyntaxHighlightHTML("<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.01 Transitional//EN\" \"http://www.w3.org/TR/html4/loose.dtd\">"));
    result.push(frontend_dumpSyntaxHighlightHTML("<!--div><div foobar-->"));
    result.push(frontend_dumpSyntaxHighlightHTML("<script></" + "script><!--div-->"));
    result.push(frontend_dumpSyntaxHighlightHTML("<script type=\"text/javascript\">document.write('<script type=\"text/javascript\"></' + 'script>');</" + "script>"));
    return result;
}

</script>
</head>

<body onload="onload()">
<p>
Tests that SourceHTMLTokenizer detects the tokens.
</p>

</body>
</html>
