LayoutTests/inspector/syntax-highlight-html-expected.txt

Tests that SourceHTMLTokenizer detects the tokens.

<html>: webkit-html-tag,webkit-html-tag,webkit-html-tag
<table cellspacing=0>: webkit-html-tag,webkit-html-tag,*,webkit-html-attribute-name,*,webkit-html-attribute-value,webkit-html-tag
<input checked value="foo">: webkit-html-tag,webkit-html-tag,*,webkit-html-attribute-name,*,webkit-html-attribute-name,*,webkit-html-attribute-value,webkit-html-tag
<table cellspacing="0" cellpadding='0'>: webkit-html-tag,webkit-html-tag,*,webkit-html-attribute-name,*,webkit-html-attribute-value,*,webkit-html-attribute-name,*,webkit-html-attribute-value,webkit-html-tag
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">: webkit-html-doctype,webkit-html-doctype,webkit-html-doctype
<!--div><div foobar-->: webkit-html-comment
<script></script><!--div-->: webkit-html-tag,webkit-html-tag,webkit-html-tag,webkit-html-tag,webkit-html-comment
<script type="text/javascript">document.write('<script type="text/javascript"></' + 'script>');</script>: webkit-html-tag,*,webkit-html-attribute-name,*,webkit-html-attribute-value,webkit-html-tag,webkit-javascript-ident,*,webkit-javascript-ident,webkit-brace-start,webkit-javascript-string,*,webkit-javascript-string,webkit-brace-end,*,webkit-html-tag,webkit-html-tag
